requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
python-binance==1.0.19
ta==0.10.2
python-dotenv==1.0.0
ccxt==4.1.87
watchdog==3.0.0 